Inherit default tick style from `matplotlib.rcParams`
Note that Matplotlib's defaults dictionary contains default settings
only for `xtick.*` and `ytick.*`. Because there is not a fixed mapping
between WCS axes and Matplotlib `x` and `y` axes, the best we can do is
arbitrarily pick one axis from which to inherit defaults.

This is explained in the docstring.